requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.4.0
bcrypt==4.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.104.1
frozenlist==1.8.0
h11==0.16.0
httptools==0.7.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.2
MarkupSafe==3.0.3
multidict==6.7.0
passlib==1.7.4
pillow==12.0.0
propcache==0.4.1
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pyotp==2.9.0
python-dotenv==1.2.1
python-multipart==0.0.6
PyYAML==6.0.3
qrcode==8.2
requests==2.32.5
sniffio==1.3.1
starlette==0.27.0
twilio==9.8.7
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
